+++ custom_components/aarlo/binary_sensor.py
@@ -14,7 +14,7 @@ from homeassistant.core import callback
 from homeassistant.helpers.config_validation import PLATFORM_SCHEMA
 
 from . import COMPONENT_ATTRIBUTION, COMPONENT_BRAND, COMPONENT_DATA, COMPONENT_DOMAIN
-from .pyaarlo.constant import (
+from pyaarlo.constant import (
     AUDIO_DETECTED_KEY,
     BUTTON_PRESSED_KEY,
     CONNECTION_KEY,
