+++ custom_components/aarlo/manifest.json
@@ -5,7 +5,7 @@
   "issue_tracker": "https://github.com/twrecked/hass-aarlo/issues",
   "dependencies": ["ffmpeg"],
   "codeowners": ["@twrecked"],
-  "requirements": ["unidecode","cloudscraper"],
+  "requirements": ["unidecode","cloudscraper>=1.2.58","pyaarlo>=0.7.1b13"],
   "version": "0.7.1b14",
   "iot_class": "cloud_push"
 }
