+++ custom_components/aarlo/sensor.py
@@ -21,7 +21,7 @@ from homeassistant.helpers.entity import Entity
 from homeassistant.helpers.icon import icon_for_battery_level
 
 from . import COMPONENT_ATTRIBUTION, COMPONENT_BRAND, COMPONENT_DATA, COMPONENT_DOMAIN
-from .pyaarlo.constant import (
+from pyaarlo.constant import (
     AIR_QUALITY_KEY,
     BATTERY_KEY,
     CAPTURED_TODAY_KEY,
